requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
async-timeout==5.0.1
audeer==2.2.1
audiofile==1.5.1
audmath==1.4.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
decorator==4.4.2
exceptiongroup==1.2.2
ffmpeg==1.4
filelock==3.17.0
flatbuffers==25.1.24
fsspec==2025.2.0
-e ./GigaAM
gigachat==0.1.38
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.34
langchain-gigachat==0.3.4
langsmith==0.3.7
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.17.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pillow==10.4.0
proglog==0.1.10
protobuf==5.29.3
pyaes==1.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
sympy==1.13.1
tenacity==9.0.0
TgCrypto==1.2.5
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
triton==3.1.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
whisper==1.1.10
zstandard==0.23.0
